fix build on windows with VS2019

+++ native/src/snappyclient/cpp/types/Timestamp.cpp
@@ -114,8 +114,11 @@ Timestamp Timestamp::parseString(const std::string& str,
         boost::date_time::parse_delimited_time_duration<
             _snappy_impl::nano_time_duration>(timeStr);
 	// nanoseconds will lie within int32 limits
-	return Timestamp(ymd.year, ymd.month, ymd.day, td.hours(), td.minutes(),
-        td.seconds(), static_cast<int32_t>(td.fractional_seconds()), utc);
+	return Timestamp(ymd.year, ymd.month, ymd.day,
+        boost::numeric_cast<uint16_t>(td.hours()),
+        boost::numeric_cast<uint16_t>(td.minutes()),
+        boost::numeric_cast<uint16_t>(td.seconds()),
+        boost::numeric_cast<int32_t>(td.fractional_seconds()), utc);
   } catch (const std::exception& e) {
     std::string err(str);
     err.append(": ").append(e.what());
@@ -139,8 +142,10 @@ std::string& Timestamp::toString(std::string& str, const bool utc) const {
       boost::posix_time::time_duration td = dateTime.time_of_day();
 
       return DateTime::toString(uint16_t(ymd.year), ymd.month.as_number(),
-          ymd.day.as_number(), td.hours(), td.minutes(), td.seconds(), m_nanos,
-          str);
+          ymd.day.as_number(),
+          boost::numeric_cast<uint16_t>(td.hours()),
+          boost::numeric_cast<uint16_t>(td.minutes()),
+          boost::numeric_cast<uint16_t>(td.seconds()), m_nanos, str);
     } catch (const std::exception&) {
       throw GET_SQLEXCEPTION2(SQLStateMessage::LANG_DATE_RANGE_EXCEPTION_MSG1,
           secsSinceEpoch);
@@ -186,7 +191,10 @@ std::ostream& operator <<(std::ostream& stream, Timestamp ts) {
       boost::posix_time::time_duration td = dateTime.time_of_day();
 
       return DateTime::toString(uint16_t(ymd.year), ymd.month.as_number(),
-          ymd.day.as_number(), td.hours(), td.minutes(), td.seconds(),
+          ymd.day.as_number(),
+          boost::numeric_cast<uint16_t>(td.hours()),
+          boost::numeric_cast<uint16_t>(td.minutes()),
+          boost::numeric_cast<uint16_t>(td.seconds()),
           ts.getNanos(), stream);
     } catch (const std::exception&) {
       throw GET_SQLEXCEPTION2(SQLStateMessage::LANG_DATE_RANGE_EXCEPTION_MSG1,
